+++ src/firmware/include/mapa.h
@@ -1,0 +1,8 @@
+#ifndef MAPA_H
+#define MAPA_H
+
+void CarregaMapa(char mapa[16][16], const char *mapa16x16);
+void IniciaMapa(char mapa[16][16]);
+void PrintaMapa(char mapa[16][16]);
+
+#endif
